+++ distribution/cauchy/doc.go
@@ -1,0 +1,9 @@
+// Package cauchy implements a cauchy continuous random variable
+//
+// Cauchy distribution is a probability distribution parameterized by
+// two parameters, namely, location and scale.
+//
+// References:
+//
+// https://en.wikipedia.org/wiki/Cauchy_distribution
+package cauchy